NASA-AMMOS/slim#46: Moved module to reserved name. ...
