+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = starterkit
+name = slim_starterkit
 url = https://github.com/NASA-AMMOS/slim-starterkit-python
 download_url = https://github.com/NASA-AMMOS/slim-starterkit-python/releases/latest
 project_urls =
@@ -38,13 +38,13 @@ zip_safe = False
 python_requires = >=3.7
 include_package_data = True
 packages =
-    starterkit
-    starterkit.api
+    slim_starterkit
+    slim_starterkit.api
 install_requires = file: requirements.txt
 
 [options.entry_points]
 console_scripts =
-    slim-starterkit-python = starterkit.hello_world:main
+    slim-starterkit-python = slim_starterkit.hello_world:main
 
 [options.packages.find]
 exclude =
