+++ slim_starterkit/api/text_processor.py
