+++ slim_starterkit/version.py
